requirements.txt
requests>=2.31.0
pyyaml>=6.0.0

# Development dependencies
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
